Add interface.tsx file and update relevant imports
