+++ components/Card/Card.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import {View, Text, StyleSheet} from 'react-native'
-import {CleanedPet} from '../../utils'
+import { CleanedPet } from '../../interfaces'
 
 type Props = {
     key: number;
