+++ utils.tsx
@@ -1,28 +1,4 @@
-interface Pet {
-    idDrink: number,
-    strDrink: string,
-    petImage: string,
-    strGlass: string,
-    dateModified: string,
-    strInstructions: string,
-    strCategory: string,
-    strCreativeCommonsConfirmed: string,
-    strDrinkThumb: string
-}
-
-export interface CleanedPet {
-    noticeID: number,
-    noticeType: string,
-    petName: string,
-    petImage: string,
-    petBreed: string,
-    dateLost: string,
-    chipNum: number,
-    petNotes: string,
-    contactNum: string,
-    contactEmail: string,
-    reward: string
-}
+import { CleanedPet, Pet } from './interfaces'
 
 const cleanData = (data: Pet[]): CleanedPet[] => {
     const cleanedData: CleanedPet[] = data.map((pet: Pet) => {
@@ -40,7 +16,6 @@ const cleanData = (data: Pet[]): CleanedPet[] => {
            reward: pet.strCreativeCommonsConfirmed
         }
     })
-  
     return cleanedData
 }
